src/requirements.txt
django 
djangorestframework 
drf-spectacular 
python-dotenv
django-jazzmin